Implements new JSON streaming and parsing.
In this commit the JSONStreamingParser is integrated  instead of the
ArduinoJSON library in order to reduce the memory consumption.
The memory consumption became a problem after the screen buffer was
doubled in size to allow for the use of Yellow.
The streaming and parsing is an adaptation of the OWM clients found in
the ThingPulse/esp8266-weather-station github project.